xen: Override sys_nanosleep with a function which doesn't quantize to HZ
The upstream NetBSD sys_nanosleep will quantize sleeping time
according to HZ; since this system default is 100, it's impossible to
even attempt to sleep for less than 10ms.

This is a hack provided by Antii that overrides the sys_nanosleep
system call with a call that will attempt to sleep for the actual
advertised nanoseconds.

Original patch by Antti Kantee <pooka@iki.fi>

Signed-off-by: George Dunlap <george.dunlap@citrix.com>